Team Generation, Random Player Gen
Random Player Generation is temporary, ideally in the future the names will depend on the country etc. That's a part of the project of its own. For now, this will temporarily do the job. Team generation is cool too. My next goal is to build and remove squads. That should finish the first user story.